feat: Add command documentation generation and update dependencies
