+++ docs/cli/witr.1
@@ -1,0 +1,86 @@
+.nh
+.TH "WITR" "1" "Jan 2026" "" ""
+
+.SH NAME
+witr - Explain processes
+
+
+.SH SYNOPSIS
+\fBwitr [process name] [flags]\fP
+
+
+.SH DESCRIPTION
+witr explains processes and their ancestry, showing how they were started and what they are doing.
+
+
+.SH OPTIONS
+\fB--env\fP[=false]
+	show only environment variables for the process
+
+.PP
+\fB-h\fP, \fB--help\fP[=false]
+	help for witr
+
+.PP
+\fB--json\fP[=false]
+	output as JSON
+
+.PP
+\fB--no-color\fP[=false]
+	disable colorized output
+
+.PP
+\fB--pid\fP=""
+	pid to look up
+
+.PP
+\fB--port\fP=""
+	port to look up
+
+.PP
+\fB--short\fP[=false]
+	short output
+
+.PP
+\fB--tree\fP[=false]
+	tree output
+
+.PP
+\fB--warnings\fP[=false]
+	show only warnings
+
+
+.SH EXAMPLE
+.EX
+
+  # Inspect a running process by name
+  witr nginx
+
+  # Look up a process by PID
+  witr --pid 1234
+
+  # Find the process listening on a specific port
+  witr --port 5432
+
+  # Show the full process ancestry (who started whom)
+  witr postgres --tree
+
+  # Show only warnings (suspicious env, arguments, parents)
+  witr docker --warnings
+
+  # Display only environment variables of the process
+  witr node --env
+
+  # Short, single-line output (useful for scripts)
+  witr sshd --short
+
+  # Disable colorized output (CI or piping)
+  witr redis --no-color
+
+  # Output machine-readable JSON
+  witr chrome --json
+
+  # Combine flags: inspect port, show ancestry, output JSON
+  witr --port 8080 --tree --json
+
+.EE
